refactor: the Student/User ListSerializer classes now share a generic ListSerializer
